Organize android
