+++ projects/Mallard/.gitignore
@@ -32,6 +32,8 @@ build/
 .gradle
 local.properties
 *.iml
+android/app/bin
+.project
 
 # Visual Studio Code
 #
