+++ projects/Mallard/android/app/src/main/java/com/guardian/editions/releasestream/ReleaseStream.java
@@ -1,4 +1,4 @@
-package com.guardian.editions;
+package com.guardian.editions.releasestream;
 
 import com.facebook.react.bridge.ReactApplicationContext;
 import com.facebook.react.bridge.ReactContextBaseJavaModule;
